// plugins/github-actions/src/api/GithubActionsApi.ts
/*
 * Copyright 2020 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { type RestEndpointMethodTypes } from "@octokit/rest";

/**
 * A client for fetching information about GitHub actions.
 *
 * @public
 */
export interface GithubActionsApi {
  listWorkflowRuns: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    pageSize?: number;
    page?: number;
    branch?: string;
  }) => Promise<
    RestEndpointMethodTypes["actions"]["listWorkflowRuns"]["response"]["data"]
  >;

  getWorkflow: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    id: number;
  }) => Promise<
    RestEndpointMethodTypes["actions"]["getWorkflow"]["response"]["data"]
  >;

  getWorkflowRun: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    id: number;
  }) => Promise<
    RestEndpointMethodTypes["actions"]["getWorkflowRun"]["response"]["data"]
  >;

  reRunWorkflow: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    runId: number;
  }) => Promise<any>;

  listJobsForWorkflowRun: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    id: number;
    pageSize?: number;
    page?: number;
  }) => Promise<
    RestEndpointMethodTypes["actions"]["listJobsForWorkflowRun"]["response"]["data"]
  >;

  downloadJobLogsForWorkflowRun: (options: {
    hostname?: string;
    owner: string;
    repo: string;
    runId: number;
  }) => Promise<
    RestEndpointMethodTypes["actions"]["downloadJobLogsForWorkflowRun"]["response"]["data"]
  >;
}